chore: resolve merge conflicts
